more RSpaceTools.mkRhoISpace to ResourceTools
