+++ rholang/src/main/scala/coop/rchain/rholang/ResourceTools.scala
@@ -2,10 +2,15 @@ package coop.rchain.rholang
 import java.io.File
 import java.nio.file.{Files, Path}
 
-import cats.effect.Resource
+import cats.effect.{Resource, Sync}
 import cats.Applicative
-import cats.effect.ExitCase.{Canceled, Completed, Error}
+import cats.effect.ExitCase.Error
 import com.typesafe.scalalogging.Logger
+import coop.rchain.models._
+import coop.rchain.rholang.interpreter.Runtime.{RhoContext, RhoISpace}
+import coop.rchain.rholang.interpreter.errors.OutOfPhlogistonsError
+import coop.rchain.rspace.history.Branch
+import coop.rchain.rspace.{Context, RSpace}
 
 import scala.reflect.io.Directory
 
@@ -25,4 +30,29 @@ object ResourceTools {
           case _ => new Directory(new File(path.toString)).deleteRecursively()
         })
     )
+
+  def mkRhoISpace[F[_]: Sync, A](
+      prefix: String = "",
+      branch: String = "test",
+      mapSize: Long = 1024L * 1024L * 4
+  ): Resource[F, RhoISpace[F]] = {
+    import coop.rchain.rholang.interpreter.storage.implicits._
+
+    def mkRspace(dbDir: Path): F[RhoISpace[F]] = {
+      val context: RhoContext = Context.create(dbDir, mapSize)
+
+      RSpace.create[
+        F,
+        Par,
+        BindPattern,
+        OutOfPhlogistonsError.type,
+        ListParWithRandom,
+        ListParWithRandomAndPhlos,
+        TaggedContinuation
+      ](context, Branch(branch))
+    }
+
+    mkTempDir(prefix)
+      .flatMap(tmpDir => Resource.liftF(mkRspace(tmpDir)))
+  }
 }
